1차적 좌석 예약 기능 완료(세부기능 제외)

+++ src/main/java/com/kh/skycastle/reservation/model/vo/Reservation.java
@@ -22,7 +22,7 @@ public class Reservation {
 	private int totalCost;
 	private Date bookEnrollDate;
 	private String status;
-	private int gradeDiscount;
+	private double gradeDiscount;
 	private Date usedDate;
 	private String startTime;
 	private String endTime;
